Switch to ZipArchive
Initially I used PharData, since it promised to support both zip
and various tar archives. But it turns out that the extracted
files are jumbled as if they were not actually decompressed.
That sounds likely as there are decompress and decompressFiles
methods on PharData. But those are mutually incompatible
(one works on zips and other on tars) and I am concerned
that it will rename the tar file unpredictably.

Let’s switch to ZipArchive, which seems to work like a charm.